+++ src/core.h
@@ -2,3 +2,6 @@
 // Author: Karol Kontny
 
 #pragma once
+
+#include <platform.h>
+#include <macros.h>
